Make Opta ID parsing sport-specific
The original Opta feed that was used was had IDs that
differ from the new one we're using. We therefore had
had to remove the prefix 't'. However, it transpired
that we should only do this removal for rugby, since
for other sports the IDs were consistent between the
two feeds.