Propagate tags from series to brands
The box-set tag was being applied to brands through the
erroneous creation of an item from the season record in MPX.
When that was removed, the box-set tag wasn't applied to brands,
since the tag is only derived from the season record.

We now propagate the same set of tags from a series to a brand
as we do from an episode to its parents.