Use equivalated brands for series match candidates
We should only consider series of equivalated brands as
candidates for matching, not the series of the candidates
of the brand match. For example:

   Ba-------Bb       Bc
    |        |       |
   Sa       Sb       Sc

If Bc was a equivalence candidate in the match of Ba, but
Bb was chosen, then only the series of Bb should be
candidates for matching to Sa. Previously this was not the
case, and we would consider series of Bb and Bc.