Only take into account next best candidate
Previously, we took the sum of all candidates with a threshold
for the strongest candidate to beat based on that. However,
with many candidates this strategy falls down. Instead, we'll
look at the second strongest candidate and ensure the
strongest candidate is at least n times greater than it.

We're only including this as part of BT VOD equivalence
in the first instance, but may roll out more widely if it
proves a good general strategy.