Containers must require a title score, not child score
BUGFIX: A container shouldn't require a child item
equivalence score to be matched; instead, they should
require a title score.

There may be no child matches, but if there is a title
match thay should suffice. In cases where there
is no schedule matching, requiring a child match before
a container match is allowed leads to a circular
dependency resulting in a failure to match ever.