Require masterAgreementServiceType OTG
Ingest rules did not correctly specify when content is available
to play. The presence of the masterAgreementServiceType:OTG is
also required (i.e. serviceType:OTG AND
masterAgreementServiceType:OTG required). This means only content
with both these tags are playable on TVE (they may still be present
in the feed though because they are playable on a STB).