Handle availabilities and versions disappearing in Nitro.
In theory, these things should never just magically vanish, they should only ever be revoked. In practice,
theory and practice aren't the same, and we need to gracefully cope with this case to then propagate these
changes to our other partners. This tries to fuzzy match versions and availabilities and revokes them
if it can't find them. This aims for recall over precision, so we *might* mistakenly revoke some
availabilities when we mismatch, but we should always catch any actual deletions.