Adding Columbus Telescope Report Handler for sending more data to Columbus Telescope
Excluding slf4j from pom as this is causing an error on start up.

Adding logging to see if equiv result is being sent to CT

Adding handlers to all Item equiv updaters to test if it's possibe to post more data to Columbus telescope.

Using prod columbus telescope as there are issues with stage CT

changing Guava version from 16 to 19