Fix nitro ingest failing when failing to resolve channel (#1128)
* Fix nitro ingest failing when failing to resolve channel

Unresolved channels are now filtered out and logged as errors
instead of the ingest completely failing to initialise.

The hardcoded list should never have anything that isn't present
in the db so unresolved channels are a potential issue. This shouldn't
be ignored but also should not stop the ingest from initialising at all.