Fix nitro ingest failing when failing to resolve channel (#1131)
Unresolved channels are now filtered out and logged as errors
instead of the ingest completely failing to initialise.

The hardcoded list should never have anything that isn't present
in the db so unresolved channels are a potential issue. This shouldn't
be ignored but also should not completely stop the ingest from
ever initialising.