Remove item from Equiv set when un-publishing
By not removing an unpublished item from the equiv set we introduce
equivalence inconsistencies. For example is A is equivalent to B and
B to C and we un-publish B, A and C would still be equivalent when
they should not. This is now resolved by removing the unpublished item
from the equiv set.