Fix start and end dates not being ingested on PA channels (#1255)
New logic was added in to update images on channels which caused the start
and end dates for those images to be lost as they were being set on the wrong
level. This retains the dates and keeps the existing specification of not
overwriting existing images if not found in new images.