Add an exact match check to equiv TitleSubset scorer. (#1260)
* Add an exact match check to equiv TitleSubset scorer.

If there is a name consisting of purely common words, or stripped
characters then titles won't match. In this case if there's an exact
match it would make sense to return a positive score.
This scorer also originally replaced an exact match scorer, so it
simply makes it a complete replacement/improvement of the old scorer.