Update bt channel data (#1262)
This creates a separate channel for each BT environment with a URI dependant on the publisher in the format "http://:key/:id" (where key is publisher key and id is the encoded atlas ID of the base channel.) This will allow channel equivalence to match them to the base channel from the URI.

In this current state, the base channels are updated in the same way as before but this can be removed later once it's confirmed working so the ingester only updates the channels for the BT environment it's running in. This also means that each environment ingester doesn't need a channel lock.