Amazon equiv should now not use the rootEquivUpdater.
The rationale is that when the ingest comes, it will update every active brand because it added a children. The delta equiv is then going to try and rerun all the brands, but we don't want to rerun all their children as well. We want equiv to rerun what we told it to rerun.