Made auditing disabled by default.
This was because nobody knew how this was used, and it is believed to
be one of the stressing factors of the database. Also it does not
seem to be intended to be on by default, and the way it works is
also dodgy, in the sense of counting what actually happens, since
some pods are configured to count and some are not.