Updated Amazon Equiv logic
Removed all uses of the ExactTitleGenerator since it
is being replaced by the AmazonTitleGenerator.

Made the AmazonTitleGenerator have a flag to
only include top level content. This is because
we do not want to blindly generate episodes with
the same title, since we may end up scoring and equiving
two episodes with the same title but from completely
different shows.
We leave the ContainerCandidates generator to
handle the cases of episodes with similar titles from
containers which have been equived together to deal with
this type of scenario.