Made amazon film year scoring done in a scorer instead of the film generator
This is because we are generating film candidates from the
AmazonTitleGenerator as well and need to score those with
the same logic.

Added a new constructor for FilmEquivalenceGenerator which
allows passing in which configurable Score values. Similarly
updated the FilmTitleMatcher to allow configurable Score values.

Changed FilmTitleMatcher to return a Score object instead of
a double to allow for returning nullScores.

Changed ItemYearScorer to allow for configurable Score
values.

Added a FilmYearScorer which extends ItemYearScorer
which will Score null for content which are not
films.