Exclude BARB broadcast equiv if BCID equiv has occurred
Since CMS and Txlog broadcasts have been found to be
inaccurate in some cases we want to prevent a txlog
equiving on bcid to one episode of CMS content
but on broadcast to another episode of CMS content
(generally from the same brand) for example.

As a result a MultiStageAllOverOrEqThresholdExtractor
was created which will extract only scores above
10 if any exist (e.g. bcid equiv) otherwise
it will extract scores above 4 (which will cover
broadcast equiv).

An unfinished inefficient generator for barb was
also removed since it will never be used.