Updated barb equiv config
Removed txlog and cdmf publishers from CMS equiv config -
This is for performance improvements and to reduce
complexity due to txlog equiv being more complicated.

Removed txlog publisher from CDMF equiv config -
Txlogs will equiv to CDMF undirectionally so
do not need to be included in the CDMF config.

Added CDMF publisher to CDMF equiv config -
CDMF should be able to equiv to other cdmf
content through oobcids/parentVersionBcids.

Removed container equiv config for txlog and cdmf publishers
since they do not have containers

Updated BarbItemUpdaterProvider to use the original
extractor but contained within an ExcludePublisher
extractor to exclude txlogs. This is in the event
that if txlogs are added back to some of the other
publishers' equiv we will avoid accidentally introducing
the broadcast/bcid equiv chaining issue that was fixed
in TxlogsItemUpdaterProvider.