Disable Five & ITV Interlinking equiv, and Itunes EPF ingest
After speaking with Karhik as part of some analysis on what we
are currently doing for Radio Times, there were a few discoveries
as outlined below. All the below decisions are part of a cost
reduction effort. It is possible that these might cause issues and
they will have to be reenabled.

a) The ingest for FIVE.TV (which seems to be C5 ondemand platform),
   has been disabled for 2 years now, so running equivalence on it
   was deemed pointless, and is now being disabled. It is possible
   we will have to fix/reenable the five.tv ingest, when the
   commercials clear out.
b) ITV Interlinking equivalence seems to be working fine, but since
   RT should be getting data from the new ITV (which is not working
   yet because ITV has not sent us a contract for 6 months now),
   whatever was there before is not required. We do not know what
   ITV Interlinking does, but on the above premise Karthik asked to
   disable it.
   The same stands for ITV Interlinking ingest. According to
   confluence, this seems to be to very old ingest anyway, already
   replaced by another ITV ingest.
c) Itunes EPF ingest was failing, and since nobody has complained
   thus far, we disabled it instead of fixing it.