Added new equiv for BBC_NITRO <-> BARB_TRANSMISSIONS
This is for work to improve equiv for BBC and to aid
in work to include Episode and Original Version PIDs
into the CCIDSTxLogs.

* Created a FirstMatchingPredicateContentEquivalenceResultUpdater
class which takes a list of EquivalenceResultUpdaterProviders
and gets an equivalence result from each in order, returning early
if a given predicate is met. This allows for additional efficiency
when running equiv since equiv logic such as generating additional
candidates can be avoided if desired.

* Added a new EquivGeneratorAndScorer which is similar to
the normal Broadcast matching equivalence but specifically
for Nitro and Txlogs which compares the actual transmission
times from Nitro with the broadcast times of txlogs
which should effectively be the same. A one second tolerance
is allowed due to Nitro having millisecond level precision
unlike txlogs.

* The Nitro <-> Txlog equiv has been changed to use this new
predicate system by first using normal BARB alias equiv,
then normal BARB broadcast matching equiv, then the new
actual transmission matching equiv, stopping early
if any strong candidates are found.