Api endpoints to update explicit equiv
ENG-282 contains the context but is to allow
for editors to explicitly equiv brands together
in order to move away from the frankly terrible
automatic brand equiv logic that happens at the
moment between most major broadcasters sources.

* Endpoint to explicitly equiv two pieces of content

* Endpoint to remove explicit equiv between two pieces of content

* Endpoint to update explicit equiv on a single piece of content
This endpoint is not needed but added due to there being a potential
need for it based on previous experience. It was also an initial
implementation for the above before deciding against using it.
Since it does not update the content record explicit equivalence
lists of other content it is explicitly equiving to it may cause
undesirable behaviour when updating explicit equivalence of other
content where the link might break due to a new link being made
to an additional piece of content from the same source - this
is due to the somewhat quirky relationship of the content record
explicit equivalences being a sort-of-subset of the lookup table
explicit equivalences.