Amazon equiv mongo title index (#1386)
Update Amazon equiv to use a separate title index instead of 
Owl search in order to get more reliable results (ENG-144)
On ingest we create an index of all Amazon content which shares the same title.

* Updated Amazon equiv to generate from a dedicated index db
* Updated Amazon ingest to populate this index
* Remove existing logic to hit owlIndexer endpoint directly

Removed all uses of the ExactTitleGenerator since it
is being replaced by the AmazonTitleGenerator.

Made the AmazonTitleGenerator have a flag to
only include top level content. This is because
we do not want to blindly generate episodes with
the same title, since we may end up scoring and equiving
two episodes with the same title but from completely
different shows.
We leave the ContainerCandidates generator to
handle the cases of episodes with similar titles from
containers which have been equived together to deal with
this type of scenario.

* Made amazon film year scoring done in a scorer instead of the film generator

This is because we are generating film candidates from the
AmazonTitleGenerator as well and need to score those with
the same logic.

Added a new constructor for FilmEquivalenceGenerator which
allows passing in which configurable Score values. Similarly
updated the FilmTitleMatcher to allow configurable Score values.

Changed FilmTitleMatcher to return a Score object instead of
a double to allow for returning nullScores.

Changed ItemYearScorer to allow for configurable Score
values.

Added a FilmYearScorer which extends ItemYearScorer
which will Score null for content which are not
films.