Added a 5min minimum brodcast duration filter to regular nitro<->txlog broadcast equiv
This is to prevent short children shows being shown back-to-back
causing equiv issues where the txlog broadcasts have shifted by
a small amount, leading to two pieces of content being deemed
valid candidates for the same broadcast slot.

This may need improvement in future but for now this is to
prevent bad equiv even at the cost of no equiv.