Rework of BarbBroadcastMatchingItemEquiv.. (Nitro<->Txlogs only atm)
The implementation looks to deal with schedule offsets between
txlogs and other schedule data, due to txlog having actual transmission
times for their broadcasts. The code looks to identify the closest
block of broadcasts for a particular show (determined by the title
scorer) on each schedule, determine which entry within the block
the subject is, and takes the corresponding entry in the other
schedule to equiv to.