Change similarEpisode logic in Barb Broadcast Equiv (#1422)
This is to fix an issue where multiple Nitro episodes
appeared one after the other with episode titles
which were just the date of transmission. Although
these were episodes of different brands they were
considered the same for block identification
since the episode titles were the same.

This logic has been changed to first compare
if the brands of the episodes are the same
before checking titles.