ENG-633 follow up - increase bbc regional txlog flexibility
5 seconds should cover a few more cases whilst still being
safer than what was originally 30 seconds.

I've decided to remove this logic from barb txlog equiv though
since it has a risk of joining two different episodes of
the same show together within the BARB view of data.

By keeping this logic only for Layer3 any episodes which
are incorrectly equived will surface in BARB as a different
BCID being listed next to a particular txlog entry.

Whilst this will still lead to incorrect equiv due to
then equiving to the other Nitro episode due to the
broadcast, in future we might be able to avoid this
with unidirectional equiv, where we won't equiv
on broadcast if we already equiv on BCID.