Copied classes from Deer related to Elastic Search
Simplified a few classes to avoid having to import
other classes that would not be needed.

Although not ideal we are copying these so that we
do not have to import deer libraries. The main reason
is that some beans in deer are being discovered
by autowiring that are causing issues sometimes when
deploying owl-api.

Since we are likely moving away from the current
elastic search anyway, a more proper solution will
likely not be worth the trouble.