(ENG-761) Remove old equiv symmetry hack
This seems to have been in place to prevent
equivalence being broken on links
which had recently been known to not be
outdated, back when equivlance was not
asymmetric.

Now that equivlance is asymmetric we
can remove this logic since such links
will not be removed anyway since
they will be kept as incoming links.