Ignore presentation channel when not in channel 4 channel map
An investigation prompted by ENG-913 found that the C4 channel map currently contains entries for C4, M4, F4, E4, 4M, and 4S, but nothing for ALL4. This causes brands with presentation channels not in the map (e.g. Community of ALL4) to not get ingested, without manually hitting a channel 4 reingest endpoint. Adding this ignore step allows these brands to be ingested, though also adds a log line for when we come across a presentation channel that does not exist in the map.