Allows BT channels to equiv to YV channels as well
It is far from clear that the current channel equivalence structure is any good.
The verifyChannelsPublishers is potentially missplaced. Since matchers can be
specific maybe the targetSources should be hardcoded in the matcher and not
harcoded on initialisation config (ChannelEquivModule). This way all relevant
BT criteria would be in a single place, the BtChannelMatcher, and the
ForcedEquivalenceChannelMatcher would not have to have an empty list of
target sources which results in a logical error and is definatelly a code smell.

I had no time to work on the structure atm, but as we seem to be getting more
channels this will likely come up again and get reworked in its own ticket.