Fix youview schedule ingest to work with new UUID's
There was a deprecated check which took the service id
from the broadcast object and ensured it is present
in the channels cast (for yv schedule ingest).
This was erroring once the switch of youview ids from
numbers to uuid's. The check was pointless at this point
so it was simply removed.