Fix youview schedule ingest to work with new UUID's (#1490)
There was a deprecated check which took the service id
from the broadcast object and ensured it is present
in the channels cast (for yv schedule ingest).
This was erroring once the switch of youview ids from
numbers to uuid's. The check was unnecessary so it was
simply removed, and getServiceId now returns string.