Lower jetty version (#1492)
A more recent version of jetty causes issues with a recent commit. It would seem it causes a different version of guava to possibly be used preventing the project from running successfully. We should investigate exactly what causes this since this version of jetty was working for a while with the project; but in the meantime we will revert it so that the recent commit will work.

The spring dependency was added back on master just as a quick test to see if that was the cause, and so has been removed again.