Sets complex broadcast times to UTC (ENG-998)
Not setting it seems to default to UTC but in a different chronology
resulting in issues when comparing it to broadcast assertions.

At the same time, we understand that Date object of the simple model
is already always in UTC, so this is expected to have no impact other
than modernise the chronology.